# __info__/swagger.yml
swagger: "2.0"
info:
  description: "DOCUMENTACION SERVICIOS REST AUDITORIA-CART-BACK"
  version: "1.1.0"
  title: "WS AUDITORIA-CART-BACK API"
  contact:
    name: "Fran Perez"
    email: "franperez.orion@gmail.com"
host: "localhost:80"
basePath: "/api/v1/auditoria"
...
securityDefinitions:
  APPKeyHeader:
    # app_key: YXBwX2JsdWVfeW9uZGVyXzIwMjA=$h
    type: "apiKey"
    in: "header"
    name: "APP_KEY"
    description: "Todas las peticiones incluirá el APP_KEY header requerido para el ID de la aplicación"
  AutorizationHeader:
    # token user: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJkbmkiOiIxMjM0NTY3OCIsIm5vbWJyZSI6IkdFTkVSSUNPIiwiYXBlcGF0IjoiR0VORVJJQ08iLCJhcGVtYXQiOiJHRU5FUklDTyIsInRpbWVTdGFtcCI6MTYwNjc1MTI2MSwiaWF0IjoxNjA2NzUxMjYxfQ.ynoTh9-KiLGtSRomfIkepCCr0QZ-VkdVyiliDYtmGY4
    type: "apiKey"
    in: "header"
    name: "AUTHORIZATION"
    description: "Sólo peticiones que necesiten validación del token"
# Apply the "app_key" header globally to all paths and operations
security:
  - APPKeyHeader: []
  - AutorizationHeader: []
paths:
...
  /box/get-container:
    get:
      tags:
        - "cartones"
      summary: "Obtener datos de cartones"
      operationId: "GetContainerOBLPN"
      parameters:
        - name: "container_id"
          in: "query"
          required: true
          type: "string"
          description: "Código de barras del contenedor"
        - name: "facility_code"
          in: "query"
          required: true
          type: "string"
          description: "Código del pallet"
        - name: "company_code"
          in: "query"
          required: true
          type: "string"
          description: "Código de la empresa asociada"
      responses:
        "200":
          description: "Consulta exitosa"
          schema:
            allOf:
              - $ref: '#/definitions/APIResponse'
              - properties:
                  Resp:
                    type: object
                    properties:
                      data:
                        type: "array"
                        items:
                          allOf:
                            - properties:
                                data:
                                  type: object
  /box/get-inventary:
    get:
      tags:
        - "cartones"
      summary: "Obtener datos de inventario de SKU"
      operationId: "GetInventaryBox"
      parameters:
        - name: "container_id"
          in: "query"
          required: true
          type: "string"
          description: "Código de barras del contenedor"
        - name: "facility_code"
          in: "query"
          required: true
          type: "string"
          description: "Código del pallet"
      responses:
        "200":
          description: "Consulta exitosa"
          schema:
            allOf:
              - $ref: '#/definitions/APIResponse'
              - properties:
                  Resp:
                    type: object
                    properties:
                      data:
                        type: "array"
                        items:
                          allOf:
                            - properties:
                                data:
                                  type: object
definitions:
  APIResponse:
    type: object
    properties:
      status:
        description: Response status `Success` or `Failure`
        type: string
        enum: ["Success", "Failure"]
      Resp:
        type: "object"
        properties:
          Error:
            type: string
          message:
            type: string
            example: Success Response
          data:
            description: Result body
      timestamp:
        description: UTC timestamp in ISO 8601 format
        type: string
    required:
      - status
      - timestamp
      - Resp
    xml:
      name: "APIResponse"
...